EditorUI moved from core to ui
