+++ packages/ckeditor5-core/src/index.ts
@@ -17,7 +17,6 @@ export { default as Context } from './context';
 export { default as ContextPlugin } from './contextplugin';
 
 export { default as Editor, type EditorReadyEvent } from './editor/editor';
-export { default as EditorUI, type EditorUIReadyEvent, type EditorUIUpdateEvent } from './editor/editorui';
 export type { EditorConfig, ToolbarConfig, ToolbarConfigItem } from './editor/editorconfig';
 
 export { default as attachToForm } from './editor/utils/attachtoform';
