+++ packages/ckeditor5-core/tests/_utils-tests/classictesteditor.js
@@ -12,7 +12,7 @@ import Plugin from '../../src/plugin';
 import HtmlDataProcessor from '@ckeditor/ckeditor5-engine/src/dataprocessor/htmldataprocessor';
 import Paragraph from '@ckeditor/ckeditor5-paragraph/src/paragraph';
 
-import EditorUI from '../../src/editor/editorui';
+import EditorUI from '@ckeditor/ckeditor5-ui/src/editorui/editorui';
 import BoxedEditorUIView from '@ckeditor/ckeditor5-ui/src/editorui/boxed/boxededitoruiview';
 import InlineEditableUIView from '@ckeditor/ckeditor5-ui/src/editableui/inline/inlineeditableuiview';
