+++ packages/ckeditor5-editor-classic/src/classiceditorui.ts
@@ -7,8 +7,8 @@
  * @module editor-classic/classiceditorui
  */
 
-import { EditorUI, type Editor, type ElementApi, type EditorUIReadyEvent } from 'ckeditor5/src/core';
-import { normalizeToolbarConfig } from 'ckeditor5/src/ui';
+import type { Editor, ElementApi } from 'ckeditor5/src/core';
+import { EditorUI, normalizeToolbarConfig, type EditorUIReadyEvent } from 'ckeditor5/src/ui';
 import { enablePlaceholder } from 'ckeditor5/src/engine';
 import { ElementReplacer } from 'ckeditor5/src/utils';
 import type ClassicEditorUIView from './classiceditoruiview';
