+++ packages/ckeditor5-editor-inline/tests/inlineeditorui.js
@@ -8,7 +8,7 @@
 import View from '@ckeditor/ckeditor5-ui/src/view';
 
 import InlineEditorUI from '../src/inlineeditorui';
-import EditorUI from '@ckeditor/ckeditor5-core/src/editor/editorui';
+import EditorUI from '@ckeditor/ckeditor5-ui/src/editorui/editorui';
 import InlineEditorUIView from '../src/inlineeditoruiview';
 import InlineEditor from '../src/inlineeditor';
 import Paragraph from '@ckeditor/ckeditor5-paragraph/src/paragraph';
