+++ packages/ckeditor5-ui/docs/examples/custom-ui.md
@@ -18,7 +18,7 @@ The editor below runs a completely custom user interface written in [Bootstrap](
 ```js
 // Basic classes to create an editor.
 import Editor from '@ckeditor/ckeditor5-core/src/editor/editor';
-import EditorUI from '@ckeditor/ckeditor5-core/src/editor/editorui';
+import EditorUI from '@ckeditor/ckeditor5-ui/src/editorui/editorui';
 import EditorUIView from '@ckeditor/ckeditor5-ui/src/editorui/editoruiview';
 import InlineEditableUIView from '@ckeditor/ckeditor5-ui/src/editableui/inline/inlineeditableuiview';
 import ElementReplacer from '@ckeditor/ckeditor5-utils/src/elementreplacer';
