+++ packages/ckeditor5-ui/src/toolbar/balloon/balloontoolbar.ts
@@ -11,8 +11,6 @@ import {
 	Plugin,
 	type Editor,
 	type EditorReadyEvent,
-	type EditorUIReadyEvent,
-	type EditorUIUpdateEvent,
 	type PluginDependencies,
 	type ToolbarConfig
 } from '@ckeditor/ckeditor5-core';
@@ -37,6 +35,8 @@ import ContextualBalloon from '../../panel/balloon/contextualballoon';
 import ToolbarView, { type ToolbarViewGroupedItemsUpdateEvent } from '../toolbarview';
 import BalloonPanelView, { generatePositions } from '../../panel/balloon/balloonpanelview';
 import normalizeToolbarConfig from '../normalizetoolbarconfig';
+
+import type { EditorUIReadyEvent, EditorUIUpdateEvent } from '../../editorui/editorui';
 
 import { debounce, type DebouncedFunc } from 'lodash-es';
