+++ packages/ckeditor5-ui/tests/tooltip/tooltipmanager.js
@@ -5,6 +5,7 @@
 
 /* global document, MouseEvent, Event */
 
+import EditorUI from '../../src/editorui/editorui';
 import View from '../../src/view';
 import BalloonPanelView from '../../src/panel/balloon/balloonpanelview';
 import Bold from '@ckeditor/ckeditor5-basic-styles/src/bold';
@@ -15,7 +16,7 @@ import global from '@ckeditor/ckeditor5-utils/src/dom/global';
 import ClassicTestEditor from '@ckeditor/ckeditor5-core/tests/_utils/classictesteditor';
 import testUtils from '@ckeditor/ckeditor5-core/tests/_utils/utils';
 import TooltipManager from '../../src/tooltipmanager';
-import { Editor, EditorUI } from '@ckeditor/ckeditor5-core';
+import { Editor } from '@ckeditor/ckeditor5-core';
 
 describe( 'TooltipManager', () => {
 	let editor, element, tooltipManager;
